+++ frontend-project/src/app/table-data/table-data.component.ts
@@ -1,9 +1,12 @@
-import { Component, DefaultIterableDiffer, Input, OnInit, Output, EventEmitter } from '@angular/core'
-import { MatTableDataSource } from '@angular/material/table'
-import { User, UserColumns } from '../interface/userData'
-import { UserService } from '../services/user.service'
+import { Component, DefaultIterableDiffer, Input, OnInit, Output, EventEmitter, ViewChild, AfterViewInit  } from '@angular/core';
+import { MatTableDataSource } from '@angular/material/table';
+import { User, UserColumns } from '../interface/userData';
+import { UserService } from '../services/user.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { Sort } from '@angular/material/sort';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
+import { MatPaginator} from '@angular/material/paginator';
 
 @Component({
   selector: 'table-data',
@@ -14,31 +17,52 @@ export class TableDataComponent {
   displayedColumns: string[] = UserColumns.map((col) => col.key)
   columnsSchema: any = UserColumns
 
-  //Parent child interaction
+  /**
+   * get dataSource data from parent component
+   */
   @Input() dataSource = new MatTableDataSource<User>();
   @Output() reset = new EventEmitter<boolean>();
 
   valid: any = {}
   is_disabled = false;
-
   sortedData: User[];
 
-  constructor(private userService: UserService, public snackBar : MatSnackBar) {
+  // Get paginator from view
+  @ViewChild('paginator') paginator!: MatPaginator;
+
+  /**
+   * Creates an instance of table component.
+   * @param userService
+   * @param snackBar
+   * @param dialog
+   */
+  constructor(private userService: UserService, public snackBar : MatSnackBar, public dialog: MatDialog) {
     this.sortedData = this.dataSource.data.slice();
   }
 
-  ngOnInit() {
-    this.dataSource.filterPredicate = (data:
-        {name: string}, filterValue: string) =>
-        data.name.trim().toLowerCase().indexOf(filterValue) !== -1;
+  /**
+   * To declare paginator values
+   * @param {void}
+   * @return {void} returns nothing
+   */
+  ngAfterViewInit(): void {
+      this.dataSource.paginator = this.paginator;
   }
 
-  //Datatable content reset, child-parent interaction
+  /**
+   * Datatable content reset, child-parent interaction through emit
+   * @param {void} nothing
+   * @return {void} returns nothing
+   */
   resetItem() {
     this.reset.emit(true);
   }
 
-  //Add and edit table content
+  /**
+   * Add and edit table data by passing row data and shows messages
+   * @param row
+   * @return {void} returns nothing
+   */
   editRow(row: User) {
     //New entry
     if (row.id === 0) {
@@ -46,6 +70,8 @@ export class TableDataComponent {
         if(row.name == '' || row.zip == '' || row.amount == '' || row.qty == '' || row.item == '') {
             //Button disabled and showing alert message
             this.is_disabled = true;
+
+            // Showing error message
             this.snackBar.open('The required fields could not Blank!', 'Required', {
                 duration : 2000,
                 panelClass : ['mat-toolbar', 'mat-warn']
@@ -64,13 +90,17 @@ export class TableDataComponent {
                     //Adding new entry
                     this.userService.addUser(row).subscribe((response: any) => {
                       if(response.success) {
+
+                        // Showing success message
                         this.snackBar.open('Data added successfully', 'Added', {
                           duration : 2000,
                           panelClass : ['mat-toolbar', 'mat-primary']
                         });
 
+                        // Table Data reseting after success
                         this.resetItem();
                       } else {
+                        // Data add faled message showing
                         this.snackBar.open('Data add failed', 'Failed', {
                           duration : 2000,
                           panelClass : ['mat-toolbar', 'mat-warn']
@@ -78,7 +108,6 @@ export class TableDataComponent {
                       }
                         
                     })
-                    //Success message
                     
                 } else {
                     //Showing error message
@@ -94,12 +123,14 @@ export class TableDataComponent {
         //Updating existing data
         this.userService.updateUser(row).subscribe((response: any) => {
           if(response.success) {
+            // Showing update success message and data resetting
             this.snackBar.open('Data updated successfully', 'Updated', {
               duration : 2000,
               panelClass : ['mat-toolbar', 'mat-primary']
             });
             this.resetItem();
           } else {
+            // Showing update failed message
             this.snackBar.open('Data update failed', 'Failed', {
               duration : 2000,
               panelClass : ['mat-toolbar', 'mat-warn']
@@ -109,28 +140,51 @@ export class TableDataComponent {
     }
   }
 
-  //Single data delete
+  /**
+   * Single data delete and shows messages
+   * @param id
+   * @return {void} returns nothing
+   */
   removeRow(id: number) {
-    this.userService.deleteUser(id).subscribe((response: any) => {
-      if(response.success) {
-        this.dataSource.data = this.dataSource.data.filter(
-          (u: User) => u.id !== id,
-        )
-        this.snackBar.open('Data removed successfully', 'Deleted', {
-            duration : 2000,
-            panelClass : ['mat-toolbar', 'mat-primary']
-        });
-      } else {
-        this.snackBar.open('Data remove failed', 'Failed', {
-          duration : 2000,
-          panelClass : ['mat-toolbar', 'mat-warn']
-        });
-      }
-        
+    /**
+     * Shows confirmation dialog selection
+     */
+    this.dialog
+      .open(ConfirmDialogComponent)
+      .afterClosed()
+      .subscribe((confirm) => {
+        // If selected confirm
+        if (confirm) {
+          // Calls delete api
+          this.userService.deleteUser(id).subscribe((response: any) => {
+            // Showing success and failed message
+            if(response.success) {
+              this.dataSource.data = this.dataSource.data.filter(
+                (u: User) => u.id !== id,
+              )
+              this.snackBar.open('Data removed successfully', 'Deleted', {
+                  duration : 2000,
+                  panelClass : ['mat-toolbar', 'mat-primary']
+              });
+            } else {
+              this.snackBar.open('Data remove failed', 'Failed', {
+                duration : 2000,
+                panelClass : ['mat-toolbar', 'mat-warn']
+              });
+            }
+              
+          })
+        }
     })
   }
 
-  //Data validity check
+  /**
+   * Data validity check
+   * @param event
+   * @param id
+   * @param key
+   * @return {void} returns nothing
+   */
   inputHandler(e: any, id: number, key: string) {
     this.is_disabled = false;
     if (!this.valid[id]) {
@@ -139,7 +193,11 @@ export class TableDataComponent {
     this.valid[id][key] = e.target.validity.valid;
   }
 
-  //Disabling submit button
+  /**
+   * Disabling submit button
+   * @param id
+   * @returns boolean
+   */
   disableSubmit(id: number) {
     if (this.valid[id]) {
       return Object.values(this.valid[id]).some((item) => item === false)
@@ -147,14 +205,28 @@ export class TableDataComponent {
     return false
   }
 
+  /**
+   * Checks all rows are selected
+   * @return boolean
+   */
   isAllSelected() {
     return this.dataSource.data.every((item) => item.isSelected)
   }
 
+  /**
+   * Checks if any rows are selected
+   * @param {void} nothing
+   * @return boolean
+   */
   isAnySelected() {
     return this.dataSource.data.some((item) => item.isSelected)
   }
 
+  /**
+   * Selects all rows 
+   * @param event
+   * @return {void} returns nothing
+   */
   selectAll(event: any) {
     this.dataSource.data = this.dataSource.data.map((item) => ({
       ...item,
@@ -162,7 +234,10 @@ export class TableDataComponent {
     }))
   }
 
-  //Sorting data
+  /**
+   * Data gets sorted here
+   * @param sort event
+   */
   sortData(sort: Sort) {
     const data = this.dataSource.data.slice();
     if (!sort.active || sort.direction === '') {
@@ -194,15 +269,26 @@ export class TableDataComponent {
     this.dataSource.data = this.sortedData;
   }
 
-  //Searching for data
+  /**
+   * Global Search for data given
+   * @param event
+   * @return {void} returns nothing
+   */
   applyFilter(event: Event) {
+    
     const filterValue = (event.target as HTMLInputElement).value;
     this.dataSource.filter = filterValue.trim().toLowerCase();
   }
 
 }
 
-//Comparing data for search
+/**
+ * Comparing data for search
+ * @param a
+ * @param b
+ * @param isAsc
+ * @return boolean
+ */
 function compare(a: number | string, b: number | string, isAsc: boolean) {
     return (a < b ? -1 : 1) * (isAsc ? 1 : -1);
 }
